add update user endpoint
